Move triple data structure convert functions to util.py
